+++ packages/game-logic/src/balance/index.ts
@@ -1,4 +1,5 @@
 export * from "./difficulty.js";
+export * from "./easier.js";
 export * from "./maintenance.js";
 export * from "./power.js";
 export * from "./reliability.js";